refactor(app): isolate app review presentation
